Moving.
I have attached a patch that fixes the bug, but I do not
understand the consequences of it.

Its effect is to get rid of Note.Moving (well, get rid of the only place where it is set to 'true').
What was the point of this variable and why do we set it?

Interestingly, this patch also fixes bug 907827.

Signed-off-by: Bert Timmerman <bert.timmerman@xs4all.nl>