Support layer types in exports
Updates to drawing/printing to support custom layer types:

* for the GUI, every drawing layer is drawn as-is

* for exporters:
- paste, silk, and mask layers are merged into the relevent generated layers
- assy and fab layers are merged into those layers
- the above layers are not exported as "copper" layers

Note that aside from the fab layer, all other special layers need to be
added to the relevent TOP or BOTTOM layer groups in the layer groups dialog.
Special layers not assigned to a group will not be exported.
(future: export those separately)