action.c: spool back by the line to be undone.
Previously the drag anchor point during drawing a new track was
reset be the line under the crosshair. The line under the
crosshair happened to be the same line  as the one to be undone
most of the time, but not always. The comment stating that the
line search would always succeed was wrong. For example when
dragging new lines while in "names only" mode.

This should fix bug 1018103, but similar cases a few lines later
in the code still exist.